Update code formatting and fix minor bugs

+++ lib/db/delete-application.jsx
@@ -3,6 +3,8 @@ import { PrismaClient } from "@prisma/client";
 export default async function deleteApplication(id) {
   const prisma = new PrismaClient();
 
+  if (!id) throw new Error("Invalid application id");
+
   try {
     const deletedApplication = await prisma.application.delete({
       where: {
@@ -12,7 +14,6 @@ export default async function deleteApplication(id) {
 
     console.log("Application deleted:", deletedApplication);
     return deletedApplication;
-    
   } catch (error) {
     console.log("Error occurred while creating event:", error);
     throw error;
